@@ -1,13 +1,11 @@
 import 'dart:async';
 import 'dart:collection';
 import 'dart:typed_data';
-import 'package:giphy_picker/src/model/giphy_client.dart';
-import 'package:giphy_picker/src/model/giphy_preview_types.dart';
 import 'package:giphy_picker/src/model/repository.dart';
 import 'package:giphy_picker/giphy_picker.dart';
 import 'package:http/http.dart' as http;
 
-typedef Future<GiphyCollection> GetCollection(
+typedef GetCollection = Future<GiphyCollection> Function(
     GiphyClient client, int offset, int limit);
 
 /// Retrieves and caches gif collections from Giphy.
@@ -24,15 +22,21 @@ class GiphyRepository extends Repository<GiphyGif> {
   GiphyRepository({
     required String apiKey,
     required this.getCollection,
-    this.maxConcurrentPreviewLoad = 4,
+    this.maxConcurrentPreviewLoad = 8,
     int pageSize = 25,
     ErrorListener? onError,
     this.previewType,
-  }) : super(pageSize: pageSize, onError: onError) {
+  }) : super(
+          pageSize: pageSize,
+          // max number of gifs, limited by giphy api
+          maxTotalCount: 5000,
+          onError: onError,
+        ) {
     _giphyClient = GiphyClient(apiKey: apiKey, client: _client);
   }
 
   /// Retrieves specified page of gif data from Giphy.
+  @override
   Future<Page<GiphyGif>> getPage(int page) async {
     final offset = page * pageSize;
     final collection = await getCollection(_giphyClient, offset, pageSize);
@@ -89,7 +93,10 @@ class GiphyRepository extends Repository<GiphyGif> {
     }
   }
 
-  Future<Uint8List?> _loadPreviewImage(GiphyGif gif) async {
+  Future<Uint8List?> _loadPreviewImage(GiphyGif? gif) async {
+    if (gif == null) {
+      return null;
+    }
     // fallback to still image if preview is empty
     String? url;
     switch (previewType) {
@@ -118,12 +125,10 @@ class GiphyRepository extends Repository<GiphyGif> {
         url = null;
         break;
     }
-    if (url == null) {
-      url = gif.images.previewGif?.url ??
-          gif.images.fixedWidthSmallStill?.url ??
-          gif.images.fixedHeightDownsampled?.url ??
-          gif.images.original?.url;
-    }
+    url ??= gif.images.previewGif?.url ??
+        gif.images.fixedWidthSmallStill?.url ??
+        gif.images.fixedHeightDownsampled?.url ??
+        gif.images.original?.url;
 
     if (url != null) {
       return await GiphyImage.load(url, client: _client);
