@@ -1,33 +1,40 @@
 import 'dart:async';
 import 'dart:collection';
+import 'dart:math';
 
 import 'package:giphy_picker/giphy_picker.dart';
 
 /// A general-purpose repository with support for on-demand paged retrieval and caching of values of type T.
 abstract class Repository<T> {
-  final HashMap<int, T> _cache = HashMap<int, T>();
-  final Set<int> _pagesLoading = Set<int>();
-  final HashMap<int, Completer<T>> _completers = HashMap<int, Completer<T>>();
+  final HashMap<int, T?> _cache = HashMap<int, T?>();
+  final Set<int> _pagesLoading = <int>{};
+  final HashMap<int, Completer<T?>> _completers = HashMap<int, Completer<T?>>();
   final int pageSize;
+  final int maxTotalCount;
   final ErrorListener? onError;
   int _totalCount = 0;
+  int _minTotalCount;
 
-  Repository({required this.pageSize, this.onError});
+  Repository(
+      {required this.pageSize, required this.maxTotalCount, this.onError})
+      : _minTotalCount = maxTotalCount;
 
   /// The total number of values available.
   int get totalCount => _totalCount;
 
   /// Asynchronously retrieves the value at specified index. When not available in local cache
   /// the page containing the value is retrieved.
-  Future<T> get(int index) {
+
+  Future<T?> get(int index) {
+    // make sure index does not exceed minTotalCount
+    index %= _minTotalCount;
+
     // index must within bounds
     assert(index == 0 || index > 0 && index < _totalCount);
 
-    final value = _cache[index];
-
-    // value is availableÍ
-    if (value != null) {
-      return Future.value(value);
+    // consult cache
+    if (_cache.containsKey(index)) {
+      return Future.value(_cache[index]);
     }
 
     final page = index ~/ pageSize;
@@ -43,7 +50,7 @@ abstract class Repository<T> {
     // value is being retrieved
     var completer = _completers[index];
     if (completer == null) {
-      completer = Completer<T>();
+      completer = Completer<T?>();
       _completers[index] = completer;
     }
 
@@ -51,18 +58,27 @@ abstract class Repository<T> {
   }
 
   void _onPageRetrieved(Page<T> page) {
-    _pagesLoading.remove(page);
-    _totalCount = page.totalCount;
+    _pagesLoading.remove(page.page);
+
+    // limit to max total
+    _totalCount = max(_totalCount, min(maxTotalCount, page.totalCount));
+    // keep min total count, used for get(index) never exceeding minimum total count
+    if (page.totalCount > 0) {
+      _minTotalCount = min(_minTotalCount, page.totalCount);
+    }
 
     if (_totalCount == 0) {
       // complete all with null
-      _completers.values.forEach((c) => c.complete(null));
+      for (var c in _completers.values) {
+        c.complete(null);
+      }
       _completers.clear();
     } else {
-      for (var i = 0; i < page.values.length; i++) {
-        // store value
+      for (var i = 0; i < pageSize; i++) {
         final index = page.page * pageSize + i;
-        final value = page.values[i];
+
+        // cache value, use null if not found
+        final value = i < page.values.length ? page.values[i] : null;
         _cache[index] = value;
 
         // complete optional completer
@@ -72,7 +88,7 @@ abstract class Repository<T> {
     }
   }
 
-  void _onPageError(int page, Object error, StackTrace stackTrace) {
+  void _onPageError(int page, GiphyError error, StackTrace stackTrace) {
     _pagesLoading.remove(page);
 
     // complete completers of this page with an error
